Add a 'dist' target
The 'dist' target creates a .tar.xz distfile from the currently checked
out working copy, including all submodules. The version is detected
automatically from release tags (with a date and optionally a branch
name appended when HEAD is not on a release tag). The name of the
package defaults to the directory name of the working copy, but can be
overridden with PKGNAME.

It requires a POSIX shell and a few tools (automatically checked).