platform.mk: Improve shell detection
Walk the base path from 'getconf PATH' ourselves, this is more reliable
than 'command -v' output.

Also add configuration variables SH and HOSTSH allowing to override the
POSIX shell and distinguish between host and targen when cross-building.

SH defaults to /bin/sh when cross-compiling and to the detected shell
with a fallback to /bin/sh when not. HOSTSH defaults to empty when
cross-compiling and to SH when not.

Rename the default sub variable to just SH, which should expand to the
correct POSIX shell path on the target system.